+++ heimdallr/api/base.py
@@ -1,70 +1,46 @@
-from env import get_env
-from heimdallr.channel import (
-    Bark,
-    BarkMessage,
-    Chanify,
-    ChanifyMessage,
-    Channel,
-    Email,
-    EmailMessage,
-    Message,
-    PushDeer,
-    PushDeerMessage,
-    Pushover,
-    PushoverMessage,
-    WecomApp,
-    WecomMessage,
-    WecomWebhook,
-)
+import asyncio
+import inspect
+import logging
+from typing import Any, Dict
+
+from heimdallr.channel.base import Channel
+from heimdallr.channel.factory import build_message
+from heimdallr.config.config import is_debug
+from heimdallr.exception import AuthException
 from heimdallr.response import Response, success
+from heimdallr.shared.config import config
+
+logger = logging.getLogger(__name__)
 
 
-def serve(
-    channel: str, title: str = "", body: str = "", key: str = "", jump_url: str = ""
-):
-    env = get_env()
-    if env.key != "" and key != env.key:
-        return {"code": -1, "message": "key not authorized"}
-    channels = channel.split("+")
-    senders = []
-    for chan in channels:
-        message: Message
-        sender: Channel
-        match chan:
-            case "bark":
-                message = BarkMessage(title, body, jump_url)
-                sender = Bark(message)
-            case "wecom-webhook":
-                message = WecomMessage(title, body)
-                sender = WecomWebhook(message)
-            case "wecom-app":
-                message = WecomMessage(title, body)
-                sender = WecomApp(message)
-            case "pushdeer":
-                message = PushDeerMessage(title, body)
-                sender = PushDeer(message)
-            case "pushover":
-                message = PushoverMessage(title, body)
-                sender = Pushover(message)
-            case "chanify":
-                message = ChanifyMessage(title, body)
-                sender = Chanify(message)
-            case "email":
-                message = EmailMessage(title, body)
-                sender = Email(message)
-            case _:
-                return {"code": 2, "message": f"{chan} is not supported"}
-        senders.append(sender)
-
-    errors = {}
-    for sender in senders:
-        rs, msg = sender.send()
-        if not rs:
-            errors[sender.get_name()] = msg
+async def serve_channels_async(key: str, title: str = "", body: str = "", **kwargs) -> Dict[str, Any]:
+    raise_exception_on_error = kwargs.pop("raise_exception_on_error", False)
+    # log the caller of this function
+    if is_debug():
+        logger.debug(f"Serving caller: {inspect.stack()[1].function}")
+    try:
+        group = config.get_group(key)
+    except AuthException as e:
+        return Response(code=-1, message=str(e)).render()
+    logger.info(f"group: {group.name}, token: {group.token}")
+    errors: Dict[str, str] = {}
+    async with asyncio.TaskGroup() as tg:
+        for chan in group.channels:
+            tg.create_task(send_to_channel(chan, title, body, errors, kwargs))
 
     if len(errors) == 0:
         return success()
     err_msg = ""
     for err in errors.items():
         err_msg += f"{err[0]} return: {err[1]}."
-    return Response(1, err_msg).render()
+    if raise_exception_on_error:
+        raise RuntimeError(err_msg)
+    return Response(code=1, message=err_msg).render()
+
+
+async def send_to_channel(chan: Channel, title: str, body: str, errors: Dict[str, str], kwargs: Dict[str, Any]) -> None:
+    logger.info(f"channel: {chan.get_name()}, channel_type: {chan.get_type()}")
+    message = build_message(chan.get_name(), title, body, **kwargs)
+    rs, err = chan.send(message)
+    if not rs:
+        errors[chan.get_name()] = err