+++ heimdallr/channel/base.py
@@ -1,10 +1,15 @@
-from typing import Tuple
+from typing import Any, Tuple
 
 
 class Message:
-    def __init__(self, title: str, body: str):
+    def __init__(self, title: str, body: str) -> None:
         self.title = title
         self.body = body
+
+    def render_message(self) -> Any:
+        """
+        Render the message to the format that the channel can understand.
+        """
 
 
 class Channel:
@@ -12,17 +17,16 @@ class Channel:
     Base class for all channels.
     """
 
-    def __init__(self, message: Message, name: str) -> None:
+    def __init__(self, name: str, type: str) -> None:
         self.name = name
-        self.message = message
+        self.type = type
 
-    def compose_message(self) -> str:
+    def _build_channel(self) -> None:
         """
-        Compose the message to be sent.
+        Get the credential of the channel.
         """
-        return ""
 
-    def send(self) -> Tuple[bool, str]:
+    def send(self, message: Message) -> Tuple[bool, str]:
         """
         Send a message to the channel.
         """
@@ -34,8 +38,8 @@ class Channel:
         """
         return self.name
 
-    def get_credential(self) -> None:
+    def get_type(self) -> str:
         """
-        Get the credential of the channel.
+        Get the type of the channel.
         """
-        pass
+        return self.type